January release

+++ economy/schema.py
@@ -464,9 +464,14 @@ class UndoProductOrderMutation(graphene.Mutation):
         session = product_order.session
         if session.closed:
             raise SuspiciousOperation("Cannot undo a product order in a closed session")
+
+        if session.type == SociSession.Type.SOCIETETEN:
+            raise SuspiciousOperation("Cannot undo a product order in a Soci Session")
+
         account = product_order.source
-        account.add_funds(product_order.cost)
-        product_order.delete()
+        with transaction.atomic():
+            account.add_funds(product_order.cost)
+            product_order.delete()
         return UndoProductOrderMutation(found=True)
 
 
@@ -498,15 +503,15 @@ class PlaceProductOrderMutation(graphene.Mutation):
         product = SociProduct.objects.get(id=product_id)
         account = user.bank_account
         cost = product.price * order_size
-
-        account.remove_funds(cost)
-        product_order = ProductOrder.objects.create(
-            source=account,
-            product=product,
-            order_size=order_size,
-            cost=cost,
-            session=session,
-        )
+        with transaction.atomic():
+            account.remove_funds(cost)
+            product_order = ProductOrder.objects.create(
+                source=account,
+                product=product,
+                order_size=order_size,
+                cost=cost,
+                session=session,
+            )
         return PlaceProductOrderMutation(product_order=product_order)
 
 
@@ -847,11 +852,11 @@ class DeleteSociOrderSessionFoodOrderMutation(graphene.Mutation):
 
 class InviteUsersToSociOrderSessionMutation(graphene.Mutation):
     class Arguments:
-        users = graphene.List(graphene.String, required=True)
+        users = graphene.List(graphene.ID, required=True)
 
     soci_order_session = graphene.Field(SociOrderSessionNode)
 
-    @gql_login_required()
+    @gql_has_permissions("economy.change_sociordersession")
     def mutate(self, info, users, *args, **kwargs):
         from economy.utils import send_soci_order_session_invitation_email
 
@@ -891,3 +896,4 @@ class EconomyMutations(graphene.ObjectType):
         DeleteSociOrderSessionFoodOrderMutation.Field()
     )
     soci_order_session_next_status = SociOrderSessionNextStatusMutation.Field()
+    invite_users_to_order_session = InviteUsersToSociOrderSessionMutation.Field()
